+++ src/web/settings_api.h
@@ -31,4 +31,11 @@ esp_err_t device_delete_handler(httpd_req_t *req);
 
 // Creating device get route callback handler.
 esp_err_t devices_get_handler(httpd_req_t *req);
+
+// Creating phone set route callback handler.
+esp_err_t phone_set_handler(httpd_req_t *req);
+
+// Creating phone get route callback handler.
+esp_err_t phone_get_handler(httpd_req_t *req);
+
 #endif